Factor out apiutil.py, log unsupported qs args.
