+++ wow/views.py
@@ -1,12 +1,12 @@
 import csv
-import functools
 import logging
 from typing import Dict, Any, List
 from pathlib import Path
 from django.http import HttpResponse, JsonResponse, Http404
 
 from .dbutil import call_db_func, exec_db_query
-from . import csvutil
+from . import csvutil, apiutil
+from .apiutil import api, get_validated_form_data
 from .forms import PaddedBBLForm, SeparatedBBLForm
 
 
@@ -17,53 +17,26 @@ SQL_DIR = MY_DIR / 'sql'
 logger = logging.getLogger(__name__)
 
 
-class InvalidFormError(Exception):
-    def __init__(self, form):
-        self.form_errors = form.errors.get_json_data()
-
-    def as_json_response(self):
-        return JsonResponse({
-            'error': 'Bad request',
-            'validationErrors': self.form_errors,
-        }, status=400)
-
-
-def apply_cors_policy(request, response):
-    response['Access-Control-Allow-Origin'] = '*'
-    return response
-
-
-def api(fn):
+def log_unsupported_request_args(request):
     '''
-    Decorator for an API endpoint.
+    Some original API endpoints sometimes got 'houseNumber', 'street',
+    'borough' query args, in which case it would look up the BBL. This
+    new implementation of the API doesn't currently support them but
+    we do want to log anytime we happen to get requests for them, to
+    diagnose whether we need to support them.
     '''
 
-    @functools.wraps(fn)
-    def wrapper(request, *args, **kwargs):
-        request.is_api_request = True        
-        try:
-            response = fn(request, *args, **kwargs)
-        except InvalidFormError as e:
-            response = e.as_json_response()
-        return apply_cors_policy(request, response)
-
-    return wrapper
-
-
-def get_validated_form_data(form_class, data) -> Dict[str, Any]:
-    form = form_class(data)
-    if not form.is_valid():
-        raise InvalidFormError(form)
-    return form.cleaned_data
+    unsupported_args = ['houseNumber', 'street', 'borough']
+    if set(request.GET.keys()).issuperset(unsupported_args):
+        logger.error(
+            f'Request contains unsupported arguments: {", ".join(unsupported_args)}')
 
 
 @api
 def address_query(request):
+    log_unsupported_request_args(request)
     args = get_validated_form_data(SeparatedBBLForm, request.GET)
     bbl = args['borough'] + args['block'] + args['lot']
-
-    # TODO: the original API also sometimes got 'houseNumber', 'street', 'borough' query
-    # args, in which case it would look up the BBL. Not sure if we need to do that here.
 
     addrs = call_db_func('get_assoc_addrs_from_bbl', [bbl])
 
@@ -117,11 +90,7 @@ def address_indicatorhistory(request):
 
 @api
 def address_export(request):
-    # TODO: The old version of this endpoint accepted
-    # 'houseNumber', 'street', 'borough' args; we should reject those w/ 400,
-    # since it would require us to make a network request to geosearch, which
-    # we don't currently support.
-
+    log_unsupported_request_args(request)
     bbl = get_request_bbl(request)
     addrs = call_db_func('get_assoc_addrs_from_bbl', [bbl])
 
@@ -146,8 +115,8 @@ def address_export(request):
 
 
 def server_error(request):
-    if getattr(request, 'is_api_request', False) is True:
-        return apply_cors_policy(request, JsonResponse(
+    if apiutil.is_api_request(request):
+        return apiutil.apply_cors_policy(request, JsonResponse(
             {'error': 'An internal server error occurred.'},
             status=500,
         ))
